Address PR feedback: extract shared CriticMarkup utility, fix comment XML
- Extract PARA_PLACEHOLDER and preprocessCriticMarkup into src/critic-markup.ts
  to deduplicate identical copies in md-to-docx.ts and manuscript-markdown-plugin.ts
- Add xml:space="preserve" to comment <w:t> elements for correct whitespace handling
- Add PARA placeholder leakage check to addition test
