+++ src/preview/manuscript-markdown-plugin.ts
@@ -2,6 +2,7 @@ import type MarkdownIt from 'markdown-it';
 import type StateInline from 'markdown-it/lib/rules_inline/state_inline.mjs';
 import type StateBlock from 'markdown-it/lib/rules_block/state_block.mjs';
 import { VALID_COLOR_IDS, getDefaultHighlightColor } from '../highlight-colors';
+import { PARA_PLACEHOLDER, preprocessCriticMarkup } from '../critic-markup';
 
 /**
  * Defines a Manuscript Markdown pattern configuration
@@ -405,56 +406,6 @@ function parseManuscriptMarkdown(state: StateInline, silent: boolean): boolean {
   }
 
   return false;
-}
-
-// Placeholder used to preserve paragraph breaks inside CriticMarkup spans.
-// Uses Private Use Area characters to avoid markdown-it's normalize step
-// which replaces \u0000 with \uFFFD.
-const PARA_PLACEHOLDER = '\uE000PARA\uE000';
-
-/**
- * Preprocess markdown source: replace \n\n inside CriticMarkup spans with a
- * placeholder so markdown-it's block parser doesn't split them into separate
- * paragraphs.
- */
-function preprocessCriticMarkup(markdown: string): string {
-  if (!markdown.includes('{++') && !markdown.includes('{--') &&
-      !markdown.includes('{~~') && !markdown.includes('{>>') &&
-      !markdown.includes('{==')) {
-    return markdown;
-  }
-
-  const markers: Array<{ open: string; close: string }> = [
-    { open: '{++', close: '++}' },
-    { open: '{--', close: '--}' },
-    { open: '{~~', close: '~~}' },
-    { open: '{>>', close: '<<}' },
-    { open: '{==', close: '==}' },
-  ];
-
-  let result = markdown;
-  for (const { open, close } of markers) {
-    let searchFrom = 0;
-    while (true) {
-      const openIdx = result.indexOf(open, searchFrom);
-      if (openIdx === -1) break;
-      const contentStart = openIdx + open.length;
-      const closeIdx = result.indexOf(close, contentStart);
-      if (closeIdx === -1) {
-        searchFrom = contentStart;
-        continue;
-      }
-      const content = result.slice(contentStart, closeIdx);
-      if (content.includes('\n\n')) {
-        const replaced = content.replace(/\n\n/g, PARA_PLACEHOLDER);
-        result = result.slice(0, contentStart) + replaced + result.slice(closeIdx);
-        searchFrom = contentStart + replaced.length + close.length;
-      } else {
-        searchFrom = closeIdx + close.length;
-      }
-    }
-  }
-  return result;
 }
 
 /** Inline rule that converts the paragraph placeholder back into line breaks in the token stream. */
